handle redirects & better reporting

+++ main.py
@@ -2,11 +2,12 @@ import asyncio
 import os
 import traceback
 
+from scraper.models.scraping_result import ScrapingWebsiteRedirect
 from scraper.scraper import Scraper
 from scraper.models.exceptions.unreachable import UnreachableException
 from scraper.models.exceptions.parsing_error import ParsingError
 from client.jobs import get_or_wait_for_new_scraping_job
-from client.data import upload_website_data, mark_site_as_unreachable, report_scraping_issue
+from client.data import upload_website_data, report_website_status
 
 from log import logging
 from config import Config
@@ -30,21 +31,26 @@ try:
             # Check if URL is valid
 
             data = asyncio.run(run_scrape_with_timeout(job.domain))
-            logging.info(f"Scraping completed. Uploading data now")
-            asyncio.run(upload_website_data(job.id, data))
-            logging.info("Upload completed")
-            if data.image_path:
-                try:
-                    os.remove(data.image_path)
-                except:
-                    pass
+            if isinstance(data, ScrapingWebsiteRedirect):
+                logging.info(f"Reporting detected redirect to {data.destination}")
+                asyncio.run(report_website_status(job, type="redirect", payload=data.dict()))
+
+            else:
+                logging.info(f"Scraping completed. Uploading data now")
+                asyncio.run(upload_website_data(job.id, data))
+                logging.info("Upload completed")
+                if data.image_path:
+                    try:
+                        os.remove(data.image_path)
+                    except:
+                        pass
         except UnreachableException as e:
             logging.warning(f"Couldn't scrape {job.domain.domain}.{job.domain.tld} due to {e}. Marking as unreachable.")
-            asyncio.run(mark_site_as_unreachable(job))
+            asyncio.run(report_website_status(job, type="unreachable" ))
 
         except (ParsingError, TimeoutError, Exception) as e:
             logging.warning(f"Couldn't scrape {job.domain.domain}.{job.domain.tld} due to {e}. Marking as unscrapable.")
-            asyncio.run(report_scraping_issue(job))
+            asyncio.run(report_website_status(job, type="issue"))
 except Exception as e:
     logging.info(f"Terminating due to {e}.")
